requirements.txt
pandas~=3.0.2
SQLAlchemy~=2.0.49
streamlit~=1.56.0
plotly~=6.6.0
python-dotenv~=1.2.2
PyMySQL~=1.1.0
spotipy~=2.26.0